// src/interfaces/Shop.Interface.ts
export interface IShop {
  sid: string;
  deliveryType: string;
  image: string;
  name: string;
  phone: string;
  rating: number;
  status: string;
  showcase: boolean;
  address: string;
  ownerName: string;
  firmName: string;
  email: string;
  bankName: string;
  IFSC: string;
  upi: string;
  shopCategory: string;
  agentID: number;
  agentScore: number;
  commissionPercent: number;
  modeOfPayment: string;
  ownerPhone: string;
  isSingleCategoryShop:boolean;
}